Refactor scripts to bin
