+++ setup.cfg
@@ -85,7 +85,7 @@ develop =
 [options.entry_points]
 # Add here console scripts like:
 console_scripts =
-    vph5_to_vpts = vptstools.scripts.vph5_to_vpts:cli
+    vph5_to_vpts = vptstools.bin.vph5_to_vpts:cli
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
