--- src/vptstools/scripts/constants.py
@@ -1,1 +1,0 @@
-CONFIG_FILE = "config.ini"
