+++ tests/test_vph5_to_vpts.py
@@ -2,7 +2,7 @@ import os
 
 from click.testing import CliRunner
 
-from vptstools.scripts.vph5_to_vpts import cli
+from vptstools.bin.vph5_to_vpts import cli
 
 # TODO - update unit tests for CLI version
 # https://click.palletsprojects.com/en/8.1.x/testing/#testing-click-applications
